# examples/tasklist-api/src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: Camunda BPM Taskpool Tasklist
  version: 0.0.1
  description: API for the task list.
basePath: '/tasklist/rest'
...
schemes:
  - http
  - https
paths:
...
  '/task/{id}/unclaim':
    post:
      tags:
        - Task
      summary: Unclaims a task.
      operationId: unclaim
      parameters:
        - in: path
          name: id
          type: string
          description: Task id.
          required: true
        - in: header
          name: X-Current-User-ID
          description: Specifies the id of current user.
          type: string
      responses:
        '204':
          description: Sucessful operation.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '404':
          description: Task not found.
...
  '/task/{id}/undefer':
    post:
      tags:
        - Task
      summary: Undefers a task.
      operationId: undefer
      parameters:
        - in: path
          name: id
          type: string
          description: Task id.
          required: true
        - in: header
          name: X-Current-User-ID
          description: Specifies the id of current user.
          type: string
      responses:
        '204':
          description: Sucessful operation.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '404':
          description: Task not found.
  '/task/{id}/complete':
    post:
      tags:
        - Task
      summary: Completes a task.
      operationId: complete
      parameters:
        - in: path
          name: id
          type: string
          description: Task id.
          required: true
        - in: header
          name: X-Current-User-ID
          description: Specifies the id of current user.
          type: string
        - in: body
          name: payload
          description: Payload to be added to the process instance on task completion.
          schema:
            $ref: '#/definitions/Payload'
      responses:
        '204':
          description: Sucessful operation.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized.
        '404':
          description: Task not found.
...
definitions:
...
  Payload:
    type: object
    additionalProperties:
      type: object
...